One more fix

+++ Arduino/gamepadcontrol.py
@@ -6,6 +6,6 @@ joy = xbox.Joystick()
 command = b'123'
 while True:
     (x,y) = joy.leftStick()
-    command = chr(int(abs(y * 255))) + chr(int(abs(x * 255))) + chr(((y > 0)*2 + (x > 0)))
+    command = chr(int(abs(y * 255))) + chr(int(abs(x * 255))) + chr((int((y > 0)*2) + int(x > 0)))
     # ser.write(command.encode('utf-8'))
     print(command)
